gateways/rpi-python/requirements.txt
spidev == 3.2
rpi.gpio == 0.6.3
paho-mqtt == 1.4.0
# pure python version of nrf24 library
git+https://github.com/jpbarraca/pynrf24.git@5aa1cc02d3f56863eb143a7105d76dc92fd25e8b
